+++ src/components/Dropdown/Dropdown.story.js
@@ -8,6 +8,7 @@ import { withInfo } from '@storybook/addon-info';
 
 import Dropdown from './index.js';
 import Button from '../Button';
+import BlockList from '../BlockList';
 import Icon from 'react-oui-icons';
 
 const data = [
@@ -33,20 +34,24 @@ stories.add('Default', withInfo()(() => {
         buttonContent='Default Dropdown'
         width={ number('width', 300) }
         arrowIcon={ true }>
-        <Dropdown.Contents>
+        <BlockList>
           {
             data.map((item, index) => {
               return (
-                <Dropdown.ListItem key={ index }>
-                  <Dropdown.BlockLink onClick={ action('click on complex item') }>
-                    <Dropdown.BlockLinkText text={ item.title } />
-                    <Dropdown.BlockLinkSecondaryText secondaryText={ item.description } />
-                  </Dropdown.BlockLink>
-                </Dropdown.ListItem>
+                <BlockList.Category header={ item.header } key={ index }>
+                  <BlockList.Item onClick={ action('click on complex item') }>
+                    <div className="flex flex-align--center">
+                      <div className="flex--1">
+                        <div>{ item.title }</div>
+                        <div className="muted micro">{ item.description }</div>
+                      </div>
+                    </div>
+                  </BlockList.Item>
+                </BlockList.Category>
               );
             })
           }
-        </Dropdown.Contents>
+        </BlockList>
       </Dropdown>
     </Container>
   );
@@ -60,20 +65,24 @@ stories.add('Error', withInfo()(() => {
         width={ number('width', 300) }
         displayError={ true }
         arrowIcon={ true }>
-        <Dropdown.Contents>
+        <BlockList>
           {
             data.map((item, index) => {
               return (
-                <Dropdown.ListItem key={ index }>
-                  <Dropdown.BlockLink onClick={ action('click on complex item') }>
-                    <Dropdown.BlockLinkText text={ item.title } />
-                    <Dropdown.BlockLinkSecondaryText secondaryText={ item.description } />
-                  </Dropdown.BlockLink>
-                </Dropdown.ListItem>
+                <BlockList.Category header={ item.header } key={ index }>
+                  <BlockList.Item onClick={ action('click on complex item') }>
+                    <div className="flex flex-align--center">
+                      <div className="flex--1">
+                        <div>{ item.title }</div>
+                        <div className="muted micro">{ item.description }</div>
+                      </div>
+                    </div>
+                  </BlockList.Item>
+                </BlockList.Category>
               );
             })
           }
-        </Dropdown.Contents>
+        </BlockList>
       </Dropdown>
     </Container>
   );
@@ -87,20 +96,24 @@ stories.add('Icon', withInfo()(() => {
         fullWidth={ boolean('fullWidth', false) }
         buttonContent={ <div>Hamburgers <span className="push-half--left"><Icon name='hamburger' /></span></div> }
         width={ number('width', 350) }>
-        <Dropdown.Contents>
+        <BlockList>
           {
             data.map((item, index) => {
               return (
-                <Dropdown.ListItem key={ index }>
-                  <Dropdown.BlockLink onClick={ action('click on complex item') }>
-                    <Dropdown.BlockLinkText text={ item.title } />
-                    <Dropdown.BlockLinkSecondaryText secondaryText={ item.description } />
-                  </Dropdown.BlockLink>
-                </Dropdown.ListItem>
+                <BlockList.Category header={ item.header } key={ index }>
+                  <BlockList.Item onClick={ action('click on complex item') }>
+                    <div className="flex flex-align--center">
+                      <div className="flex--1">
+                        <div>{ item.title }</div>
+                        <div className="muted micro">{ item.description }</div>
+                      </div>
+                    </div>
+                  </BlockList.Item>
+                </BlockList.Category>
               );
             })
           }
-        </Dropdown.Contents>
+        </BlockList>
       </Dropdown>
     </Container>
   );
@@ -115,22 +128,24 @@ stories.add('Z-index', withInfo()(() => {
             isDisabled={ boolean('isDisabled', false) }
             buttonContent='Dropdown'
             width={ number('width', 300) }>
-            <Dropdown.Contents
-              minWidth={ 300 }
-              direction={ 'up' }>
+            <BlockList>
               {
                 data.map((item, index) => {
                   return (
-                    <Dropdown.ListItem key={ index }>
-                      <Dropdown.BlockLink onClick={ action('click on complex item') }>
-                        <Dropdown.BlockLinkText text={ item.title } />
-                        <Dropdown.BlockLinkSecondaryText secondaryText={ item.description } />
-                      </Dropdown.BlockLink>
-                    </Dropdown.ListItem>
+                    <BlockList.Category header={ item.header } key={ index }>
+                      <BlockList.Item onClick={ action('click on complex item') }>
+                        <div className="flex flex-align--center">
+                          <div className="flex--1">
+                            <div>{ item.title }</div>
+                            <div className="muted micro">{ item.description }</div>
+                          </div>
+                        </div>
+                      </BlockList.Item>
+                    </BlockList.Category>
                   );
                 })
               }
-            </Dropdown.Contents>
+            </BlockList>
           </Dropdown>
           <h1>This text should be behind the open dropdown</h1>
         </ScrollContainer>