+++ src/components/Dropdown/tests/index.js
@@ -136,14 +136,6 @@ describe('components/Dropdown', () => {
       <Dropdown
         buttonContent='Dropdown'
         displayError={ true }>
-        <ul>
-          { data.map((item, index) => {
-            return (
-              <li key={ index }>{ item.title }</li>
-            );
-          })
-          }
-        </ul>
       </Dropdown>
     );
     expect(shallowToJson(component)).toMatchSnapshot();
@@ -154,14 +146,6 @@ describe('components/Dropdown', () => {
       <Dropdown
         activator={ <button>Click me</button> }
         displayError={ true }>
-        <ul>
-          { data.map((item, index) => {
-            return (
-              <li key={ index }>{ item.title }</li>
-            );
-          })
-          }
-        </ul>
       </Dropdown>
     );
     expect(shallowToJson(component)).toMatchSnapshot();