+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+### Fixed
+- [Patch] Undo breaking changes to Dropdown, remove filter functionality from SelectDropdown that required Dropdown's breaking changes
 
 ## 31.7.1 - 2018-07-09
 ### Fixed